% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_create_env.R
\name{conda_create_walking_env}
\alias{conda_create_walking_env}
\title{Create Conda Environment for Walking}
\usage{
conda_create_walking_env(envname = "stepcount", ...)
}
\arguments{
\item{envname}{environment name}

\item{...}{additional arguments to pass to \code{\link[reticulate:conda-tools]{reticulate::conda_create()}}}
}
\value{
Output of \link[reticulate:conda-tools]{reticulate::conda_create}
}
\description{
Create Conda Environment for Walking
}
