% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_require_stepcount.R
\name{py_require_stepcount}
\alias{py_require_stepcount}
\title{Command for \code{py_require} for \code{stepcount}}
\usage{
py_require_stepcount(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[reticulate:py_require]{reticulate::py_require()}}}
}
\value{
A logical value indicating whether the package is available.
}
\description{
Command for \code{py_require} for \code{stepcount}
}
