% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_stepcount.R
\name{py_stepcount}
\alias{py_stepcount}
\title{Perform step count calculation in a separate Python environment}
\usage{
py_stepcount(
  ...,
  pyenv_function = function() {
     stepcount::py_require_stepcount()
 },
  show = TRUE
)
}
\arguments{
\item{...}{arguments passed to \code{\link[=stepcount]{stepcount()}}}

\item{pyenv_function}{function that loads the \code{stepcount} Python package.
By default, it uses \link{py_require_stepcount}.
If this function has an \code{args} argument, the output
of \code{pyenv_function} will be re-assigned to \code{args}.}

\item{show}{Logical, whether to show the standard output on the
screen while the child process is running, passed to \code{\link[callr:r]{callr::r()}}}
}
\value{
The output from \code{\link[=stepcount]{stepcount()}}.
A tibble with minute-level \code{time}, \code{steps}, and \code{walking} columns.
}
\description{
Perform step count calculation in a separate Python environment
}
\examples{
\donttest{
  file = system.file("extdata/P30_wrist100.csv.gz", package = "stepcount")
  if (stepcount_check()) {
    df = readr::read_csv(file)
    out = try({py_stepcount(df, sample_rate = 100)})
  }
}
}
