% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{sc_rename_data}
\alias{sc_rename_data}
\alias{sc_write_csv}
\title{Rename data for Stepcount}
\usage{
sc_rename_data(data)

sc_write_csv(data, path = tempfile(fileext = ".csv"))
}
\arguments{
\item{data}{a \code{data.frame} of raw accelerometry}

\item{path}{path to the CSV output file}
}
\value{
A \code{data.frame} of renamed columns
}
\description{
Rename data for Stepcount
}
