% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_stepcount.R
\name{have_stepcount}
\alias{have_stepcount}
\alias{stepcount_check}
\alias{stepcount_version}
\title{Check the \code{stepcount} Python Module}
\usage{
have_stepcount()

stepcount_check()

stepcount_version()
}
\value{
A logical value indicating whether the \code{stepcount} Python module is available.
}
\description{
Check the \code{stepcount} Python Module
}
\examples{
if (have_stepcount()) {
   stepcount_version()
}
}
