% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepcount.R
\name{stepcount}
\alias{stepcount}
\title{Run Stepcount Model on Data}
\usage{
stepcount(
  file,
  model_type = c("ssl", "rf"),
  model_path = NULL,
  pytorch_device = c("cpu", "cuda:0"),
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Run Stepcount Model on Data
}
\examples{
file = system.file("extdata/P30_wrist100.csv.gz", package = "stepcount")
if (have_stepcount() && have_stepcount_condaenv()) {
  out = stepcount(file = file)
}
}
