% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepcount.R
\name{stepcount}
\alias{stepcount}
\title{Run Stepcount Model on Data}
\usage{
stepcount(
  file,
  model_type = c("ssl", "rf"),
  model_path = NULL,
  pytorch_device = c("cpu", "cuda:0"),
  verbose = TRUE
)
}
\arguments{
\item{file}{accelerometry file to process, including CSV,
CWA, GT3X, and \code{GENEActiv} bin files}

\item{model_type}{type of the model: either random forest (rf) or
Self-Supervised Learning model (ssl)}

\item{model_path}{the file path to the model.  If on disk, this can be
re-used and not re-downloaded.  If \code{NULL}, will download to the
temporary directory}

\item{pytorch_device}{device to use for prediction for PyTorch.}

\item{verbose}{print diagnostic messages}
}
\value{
A list of the results (\code{data.frame}),
summary of the results, adjusted summary of the results, and
information about the data.
}
\description{
Run Stepcount Model on Data
}
\examples{
file = system.file("extdata/P30_wrist100.csv.gz", package = "stepcount")
df = readr::read_csv(file)
if (stepcount_check()) {
  out = stepcount(file = file)
  out_df = stepcount(file = df)
}
}
