% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_stepcount.R
\name{install_stepcount}
\alias{install_stepcount}
\alias{have_stepcount}
\alias{stepcount_check}
\alias{stepcount_version}
\title{Install the \code{stepcount} Python Module}
\usage{
install_stepcount(packages = "stepcount", ...)

have_stepcount()

stepcount_check()

stepcount_version()
}
\arguments{
\item{packages}{packages to install.
If \code{stepcount} is not included, it will be added.  This package is
known to work with \verb{stepcount==3.2.4}}

\item{...}{Additional arguments to pass to \code{\link[reticulate:py_install]{reticulate::py_install()}},
other than \code{pip} (\code{pip = TRUE} enforced)}
}
\value{
Output of \link[reticulate:py_install]{reticulate::py_install}
}
\description{
Install the \code{stepcount} Python Module
}
\examples{
if (have_stepcount()) {
   stepcount_version()
}
}
